Add `from_json_value` for `GeoJson` enum
There already exists `GeoJSON::from_json_object`, which converts a
`serde_json::Map<String, serde_json::Value>` to a geojson instance. It
would be convenient to be able to convert straight from a
`serde_json::Value`, as this is what is constructed with the `json!`
macro.

This commit implements `GeoJson::from_json_value` as a thin wrapper
around `from_json_object`.