Add `TryFrom` impls for JsonObject and JsonValue
For each GeoJson type, an `impl TryFrom<JsonObject>` and `impl
TryFrom<JsonValue>` is added, and the analagous `from_json_object` and
`from_json_value` functions are converted to use the new traits.

This makes it possible to simplify the instantiation of GeoJson types
like this:

```rust
let feature: Feature = json!({
    "type": "Feature",
    "geometry": {
        "type": "Point",
        "coordinates": [102.0, 0.5]
    },
    "properties": null,
}).try_into().unwrap();
```

impls are added for:

* `GeoJson`
* `Geometry`
* `Feature`, and
* `FeatureCollection`.

API stability is maintained - the only change is that the
`from_json_object` methods take `JsonObject` instead of
`mut JsonObject`, but this should not break anything.

Tests are added where appropriate.