feature without 'properties' key doesn't error
Even though properties is required for features, sometimes in the wild
it's missing. In the interest of being strict about what we emit while
being permissive about what we accept, I think it'd be helpful for users
to go forward when there is no properties key, and treat it just like
null or empty properties.